requirements.txt
Flask==1.1.1
Flask-Bootstrap==3.3.7.1
Flask-Cors==3.0.8
Flask-Dance==2.2.0
Flask-Login==0.4.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.0
Flask-Script==2.0.6
Flask-Sockets==0.2.1
Flask-Testing==0.7.1
Flask-WTF==0.14.2
Jinja2==2.10.1
Mako==1.1.0
MarkupSafe==1.1.1
Pillow==6.1.0
SQLAlchemy==1.3.8
WTForms==2.2.1
Werkzeug==0.15.6
alembic==1.1.0
argparse==1.2.1
backports.ssl-match-hostname==3.7.0.1
blinker==1.4
gevent==1.4.0
gevent-websocket==0.10.1
greenlet==0.4.15
gunicorn==19.9.0
httplib2==0.13.1
itsdangerous==1.1.0
nose==1.3.7
oauth2==1.9.0.post1
oauthlib==3.1.0
psycopg2-binary==2.8.3
pycrypto==2.6.1
redis==3.3.8
requests==2.22.0
requests-oauthlib==1.2.0
selenium==3.141.0
six==1.12.0
twitter==1.18.0
websocket-client==0.56.0
